feat(tunnel): add agent-hosted dynamic (-D SOCKS) forwarding
Lift the dynamic (SOCKS5) forward engine into termihub_core::tunnel so the
identical forwarder runs on the desktop or on an agent, mirroring the -L/-R
lifts. Add a TunnelForwardSpec::Dynamic variant + AgentTunnelRegistry
start_dynamic, route agent-hosted -D through the agent RPC in the desktop
start_agent_tunnel (removing the -D-errors path), and report reachableFrom
classified from the loopback-safe SOCKS bind.

Part of #2185
Closes #2198
