+++ agent/src/handler/dispatch.rs
@@ -1448,6 +1448,10 @@ fn register_tunnel_start(module: &mut RpcModule<Mutex<HandlerState>>) -> anyhow:
                 .start_remote(&p.tunnel_id, &p.ssh_config, forward)
                 .await
                 .map_err(|e| rpc_err(errors::TUNNEL_START_FAILED, e.to_string()))?,
+            TunnelForwardSpec::Dynamic(forward) => registry
+                .start_dynamic(&p.tunnel_id, &p.ssh_config, forward)
+                .await
+                .map_err(|e| rpc_err(errors::TUNNEL_START_FAILED, e.to_string()))?,
         };
 
         Ok::<_, ErrorObjectOwned>(
